Finished out shib functionality. It works for me on my test setup now, modulo having to mock in some shib attributes that I don't get.
The current_user stuff had to be redone to use the id of the user -
shib users don't have a login_id.

We also have to suspend checking the presence of the login_id when
making a shibboleth user in the User model.

Added routes needed.

We still might need to mess with the configuration of the omniauth plugin
a bit, but I really can't do that without seeing more of how the live
set up works.